Issue #3250 - Adjusts functional tests to the wizard form
some tests have been removed because duplicated elsewhere. some partially rewritten. The new form is definitely more challending to test than the previous one and there is a lot of repeat. In fact I would probably see in the future a way where we can branch scenarios in the tests through meta function, instead of repeating the same code again and again.
